Add framework.uuid_db as a JSONHashFile
